+++ Changelog.md
@@ -9,6 +9,9 @@ Riskbook took over maintainership
 + Add integration tests
 + Add basic Ejabberd api support
   This is to allow basic integration tests.
++ Add simple integration tests
++ Add show instance for SomeStanza
++ More flexible show instance for Stanza
 
 ## Version 1.0.2 - 13.12.2019